+++ QuranMajeed/Quiz/Main/ViewModel/QuizViewModel.swift
@@ -10,7 +10,6 @@ import Foundation
 import PDFKit
 import QuranKit
 import ReadingService
-import Localization
 
 extension QuizView {
     class ViewModel: ObservableObject {
@@ -94,23 +93,3 @@ extension QuizView {
         private let readingPreferences = ReadingPreferences.shared
     }
 }
-
-
-
-extension Sura {
-    public var localizedSuraNumber: String {
-        NumberFormatter.shared.format(suraNumber)
-    }
-
-    public func localizedName(withPrefix: Bool = false, withNumber: Bool = false, language: Language? = nil) -> String {
-        var suraName = l("sura_names[\(suraNumber - 1)]", table: .suras, language: language)
-        if withPrefix {
-            suraName = lFormat("quran_sura_title", table: .android, language: language, suraName)
-        }
-        if withNumber {
-            suraName = "\(localizedSuraNumber). \(suraName)"
-        }
-        return suraName
-    }
-}
-
